feat"SNG-6 Feature Movie Details": Implement movie details screen with UI components and API integration of the first endpoint - added new colors and theme - added cached_network_image

+++ lib/features/details/data/models/movie_detail_model.dart.dart
@@ -5,7 +5,7 @@ class MovieDetailsModel {
   String? backdropPath;
   String? releaseDate;
   String? overview;
-  int? voteAverage;
+  double? voteAverage;
   List<Genres>? genres;
   int? runtime;
